refactor(harvest): split services.go (1675 → 8 files) by concern
Fifth worst-first file-size refactor. Four services were sharing one file;
each now lives in its own ≤350-line same-package file plus three
GeneratorService-method files split by responsibility.

  services.go              281   GeneratorService + Generate orchestration + constants
  services_pdf.go          217   PDF report renderer + per-section helpers
  services_formats.go      130   HTML / CSV / JSON renderers
  services_reports.go      142   report-record CRUD + GenerateFromTemplate
  services_export.go       191   bulk Export + per-entity exporters + dataToCSV
  services_template.go     208   TemplateService (built-in + custom templates)
  services_scheduler.go    350   SchedulerService + calculateNextRun
  services_aggregator.go   271   AggregatorService + GetTrends

All under the 600-line warn budget.

Drive-by cleanups forced by --new-from-rev on relocated lines:

- Reused the existing statusCritical constant for the two "critical" string
  literals that goconst flagged (the moved vulnerability-severity switch
  and the moved template section ID).
- Extracted entityDevices = "devices" in services.go and threaded it
  through the Export switch, GetTrends metric switch, and inventory
  template section ID.

Build, vet, lint (--new-from-rev=HEAD~1: 0 issues), and harvest tests pass.
